Move options parsing to a separate module
This cleans up the main process, and makes options parsing easier
to test.
